Moved helper scripts to `/scripts`
